refactor(jact-cli): extract fix helpers to core/citationFixer module (closes #46)
Move applyPathConversion, applyAnchorFix, parseAvailableHeaders,
findBestHeaderMatch, urlEncodeAnchor (and private normalizeAnchorForMatching)
from JactCli class body into src/core/citationFixer.ts as exported functions.
JactCli.fix() now imports and delegates to them; HeaderObject and PathConversion
type imports removed from jact.ts as no longer needed there.
